+++ docs/CHANGE_LOG.md
@@ -1,5 +1,19 @@
 # Calendar.js - Change Log:
 
+## Version 2.3.1:
+- Added new setting "shortcutKeysEnabled" (defaults to true), which states if the shortcut keys are enabled (always off in DatePicker mode).
+- Fixed some mistakes in the documentation (CLASS.md, CUSTOM_TRIGGERS.md, FUNCTIONS.md, SEARCH_OPTIONS.md).
+- Added "Select All" and "Select None" buttons to the side menu sections (only visible when a section is open).
+- Increased the width of the side menu slight for desktop and mobile devices (to give a bit more room for the new section buttons).
+- When pressing escape, if events are copied and another view is opened, the events copied are cleared as the first action (the view will remain open).
+- Added "VALARM" support for exported iCAL events (states when an alert should be shown for the event, which defaults to the events "from" date/time).
+- Added DOCTYPE, HEAD, and TITLE support for exported HTML events, including meta for CHARSET and LAST-MODIFIED.
+- Added "Filename" and "Last Modified" support export TXT events.
+- Added a "Calendar" node to the exported XML events, which contains sub-nodes "Filename" and "LastModified".
+- Pressing enter in the "Filename" field on the Export Dialog will now trigger the "Export" button.
+
+<br>
+
 ## Version 2.3.0:
 
 #### **New Features:**